requirements.txt
web3==6.20.4
pymongo==4.15.1
requests==2.32.3
APScheduler==3.11.0
python-dotenv==1.0.1
protobuf==4.25.3
eth-account==0.9.0
eth-keys==0.4.0
eth-abi==4.2.1
eth-rlp==0.3.0
hexbytes==0.3.1
